+++ src/main/java/so/edu/uct/BloodBank/api/UserResource.java
@@ -3,9 +3,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import so.edu.uct.BloodBank.dto.LoginRequest;
+import so.edu.uct.BloodBank.model.BloodType;
 import so.edu.uct.BloodBank.model.User;
 import so.edu.uct.BloodBank.service.UserService;
 
+import java.util.HashMap;
 import java.util.List;
 import java.util.Objects;
 
@@ -26,8 +28,17 @@ public class UserResource {
     // 2. Get Specific User By ID
 
     @GetMapping(value = "/{id}")
-    public User getUser(@PathVariable Long id){
-        return userService.getUserById(id);
+    public ResponseEntity<?> getUser(@PathVariable Long id){
+        User getUser = userService.getUserById(id);
+        if (getUser == null){
+            HashMap<String, String> response = new HashMap<>();
+            // Add Keys and Values
+            response.put("status", "404");
+            response.put("message", "User Not Found");
+            System.out.println(response);
+            return ResponseEntity.status(404).body(response);
+        }
+        return ResponseEntity.ok().body(userService.getUserById(id));
     }
 
 
@@ -41,19 +52,35 @@ public class UserResource {
     // 4. Update Specific User By ID
 
     @PutMapping(value = "/{id}")
-    public User updateUser(@RequestBody User user, @PathVariable Long id) {
+    public ResponseEntity<?> updateUser(@RequestBody User user, @PathVariable Long id) {
         User updatedUser = userService.getUserById(id);
+        if (updatedUser == null){
+            HashMap<String, String> response = new HashMap<>();
+            // Add Keys and Values
+            response.put("status", "404");
+            response.put("message", "User Not Found");
+            System.out.println(response);
+            return ResponseEntity.status(404).body(response);
+        }
         updatedUser.setName(user.getName());
-        return userService.saveUser(updatedUser);
+        return ResponseEntity.ok().body(userService.saveUser(updatedUser));
     };
 
     // 5. Delete Specific User By ID
 
     @DeleteMapping(value = "/{id}")
-    public User deleteUser(@PathVariable Long id){
+    public ResponseEntity<?> deleteUser(@PathVariable Long id){
         User deletedUser = userService.getUserById(id);
+        if (deletedUser == null){
+            HashMap<String, String> response = new HashMap<>();
+            // Add Keys and Values
+            response.put("status", "404");
+            response.put("message", "User Not Found");
+            System.out.println(response);
+            return ResponseEntity.status(404).body(response);
+        }
         userService.deleteUser(id);
-        return deletedUser;
+        return ResponseEntity.ok().body(deletedUser);
     }
 
     // 6. Login
